+++ suite/admin/actions/notifications/delete.php
@@ -8,8 +8,8 @@ $base_url = $ini["general"]["base_url"];
 $id = $_GET["id"];
 $arbeit->auth()->login_validation();
 if($arbeit->benutzer()->is_admin($arbeit->benutzer()->get_user($_SESSION["username"]))){
-    if($arbeit->notifications()->delete_calendar_entry($id) == true){
-        header("Location: http://{$base_url}/suite/?info=calendar_entry_deleted");
+    if($arbeit->notifications()->delete_notifications_entry($id) == true){
+        header("Location: http://{$base_url}/suite/?info=notifications_entry_deleted");
     }   
 } else {
     header("Location: http://{$base_url}/suite/?info=noperms");