+++ suite/admin/actions/notifications/edit.php
@@ -7,8 +7,8 @@ $ini = $arbeit->get_app_ini();
 $base_url = $ini["general"]["base_url"];
 $arbeit->auth()->login_validation();
 if($arbeit->benutzer()->is_admin($arbeit->benutzer()->get_user($_SESSION["username"]))){
-    if($arbeit->notifications()->edit_calendar_entry($_GET["id"], $_POST["uhrzeit"], $_POST["datum"], $_POST["ort"], $_POST["notiz"]) == true){
-        header("Location: http://{$base_url}/suite/?info=calendar_entry_edited");
+    if($arbeit->notifications()->edit_notifications_entry($_GET["id"], $_POST["uhrzeit"], $_POST["datum"], $_POST["ort"], $_POST["notiz"]) == true){
+        header("Location: http://{$base_url}/suite/?info=notifications_entry_edited");
     }   
 } else {
     header("Location: http://{$base_url}/suite/?info=noperms");