+++ suite/admin/notifications/edit.php
@@ -13,7 +13,7 @@ if(!@$arbeit->benutzer()->is_admin($_SESSION["username"])){
     header("Location: http://{$base_url}/suite/?info=noperms");
 }
 $id = $_GET["id"];
-$data = $arbeit->notifications()->get_calendar_entry($id);
+$data = $arbeit->notifications()->get_notifications_entry($id);
 ?>
 <!DOCTYPE html>
 <html>
@@ -27,7 +27,7 @@ $data = $arbeit->notifications()->get_calendar_entry($id);
 
         <h1><?php echo $language["title"] ?></h1>
         <div class="box">
-            <form action="/suite/admin/actions/calendar/edit.php?id=<?php echo $id; ?>" method="POST">
+            <form action="/suite/admin/actions/notifications/edit.php?id=<?php echo $id; ?>" method="POST">
                 <label><?php echo $language["label_date"] ?>: </label><input type="date" name="datum" value="<?php echo $data["datum"] ?>">
                 <br>
                 <label><?php echo $language["label_time"] ?>: </label><input type="time" name="uhrzeit" value="<?php echo $data["uhrzeit"]; ?>">